Implement query logic
This is a general query module that will be used
to query both project for their todos and the
database at large for their projects. We have two
methods.

cleanAndValidateQuery(queryCriteria,
queryParams) checks if the query can be turned
into a js object, and then if the params in the js
object are included with the param list of
whatever container we are querying within.

query(items, queryParams, queryCriteria) runs
cleanAndValidateQuery on the query, and then uses
a filter to return the items in the item list that
matches that query. Items is what is stored in our
database, i.e. a list of projects, todos, etc.